Added a type cast for pandas object

+++ px4/combine_and_resample_px4_nogui.py
@@ -103,7 +103,7 @@ def combine_and_resample_px4_nogui(input_path,file_suffix=''):
     big_df['time_properformat'] = pd.to_timedelta(big_df.index/10.0**6,'S')
     
     # switch the index of the big_df to proper time delta column
-    big_df.index = pd.to_datetime(big_df.time_properformat)
+    big_df.index = pd.to_datetime(big_df.time_properformat.astype('int64'))
     
     # fix this for 250Hz rate since the auto calculator is causing issues
     min_sampletime = .004
